pyproject.toml
[project]
name = "aws-cognito-circleci"
version = "0.1.0"
description = "A Python-based tool for automating AWS Cognito User Pool setup and management through CircleCI pipelines."
readme = "README.md"
authors = [
    { name = "Benito Martin"},
]
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.29",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]

[dependency-groups]
ci = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "mypy>=1.15.0",
    "ruff>=0.11.4",
]

dev = [
    "pre-commit>=4.2.0",
    "types-boto3>=1.37.29",
    "types-requests>=2.32.0.20250328",
    "boto3-stubs>=1.37.37",
    "mypy-boto3-cognito-idp>=1.37.13.post1",
]

######################################
# --- Linting & Formatting Tools --- #
######################################

[tool.ruff]
# Assume Python 3.12
target-version = "py312"

# Same as Black.
line-length = 120
indent-width = 4

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Whether to show an enumeration of all fixed lint violations
show-fixes = true

# Enable common lint rules.
lint.select = [
  "B",    # flake8-bugbear
  "E",    # pycodestyle
  "F",    # Pyflakes
  "I",    # isort
  "SIM",  # similarity
  "UP",   # pyupgrade
]
lint.ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ruff.lint.mccabe]
# Maximum allowed McCabe complexity.
max-complexity = 10


#########################
# --- Static Typing --- #
#########################

[tool.mypy]
# Mypy configuration.
files = ["app","src"]
ignore_missing_imports = false
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

#########################
# --- Testing Tools --- #
#########################

[tool.pytest.ini_options]
testpaths = [ "tests" ]
python_files = [ "test_*.py" ]
addopts = "-ra -v -s"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
